+++ api/src/routes/product.js
@@ -12,6 +12,10 @@ router.get('/', (req, res, next) => {
 });
 
 router.post('/', async (req, res) => {
+	// Guard clauses
+	if (!req.isAuthenticated()) return res.status(401).send('No estás logueado');
+	if (req.user.rol !== 'Admin') return res.status(401).send('No eres admin');
+
 	const {name, price, stock, image, description} = req.body;
 	if (!name || !price || typeof stock !== 'number' || !image || !description)
 		return res.status(400).send('Falta algún parámetro o stock typeof incorrecto');
@@ -40,6 +44,10 @@ router.get('/:id', (req, res) => {
 });
 
 router.delete('/:id', (req, res) => {
+	// Guard clauses
+	if (!req.isAuthenticated()) return res.status(401).send('No estás logueado');
+	if (req.user.rol !== 'Admin') return res.status(401).send('No eres admin');
+
 	const {id} = req.params;
 
 	Product.destroy({where: {id}, include: [Pics]}).then(response => {
@@ -49,6 +57,9 @@ router.delete('/:id', (req, res) => {
 });
 
 router.put('/:id', async (req, res) => {
+	// Guard clauses
+	if (!req.isAuthenticated()) return res.status(401).send('No estás logueado');
+	if (req.user.rol !== 'Admin') return res.status(401).send('No eres admin');
 
 	const {name, price, stock, description, image} = req.body;
 	// const image = req.body.image ? req.body.image[0] : null;
@@ -67,15 +78,15 @@ router.put('/:id', async (req, res) => {
 		robot.price = price || robot.price;
 		robot.stock = stock || stock === 0 ? stock : robot.stock;
 		if (image) {
-		// 	await Pics.findOrCreate({where: {imageUrl: image, productId: robot.id}});
-		// 	robot.image = image;
-		await Pics.destroy({where: {productId: robot.id}});
+			// 	await Pics.findOrCreate({where: {imageUrl: image, productId: robot.id}});
+			// 	robot.image = image;
+			await Pics.destroy({where: {productId: robot.id}});
 
 			image.map(img => {
 				Pics.create({imageUrl: img})
 					.then(newPic => robot.addPic(newPic))
 					.then(response => res.status(201).send(response));
-			})
+			});
 		}
 
 		await robot.save();
@@ -87,6 +98,10 @@ router.put('/:id', async (req, res) => {
 });
 
 router.post('/:idProducto/category/:idCategoria', async (req, res) => {
+	// Guard clauses
+	if (!req.isAuthenticated()) return res.status(401).send('No estás logueado');
+	if (req.user.rol !== 'Admin') return res.status(401).send('No eres admin');
+
 	const {idProducto, idCategoria} = req.params;
 	const producto = await Product.findByPk(idProducto);
 	const categoria = await Categories.findByPk(idCategoria);
@@ -101,20 +116,15 @@ router.post('/:idProducto/category/:idCategoria', async (req, res) => {
 });
 
 router.delete('/:idProducto/category/:idCategoria', (req, res) => {
+	// Guard clauses
+	if (!req.isAuthenticated()) return res.status(401).send('No estás logueado');
+	if (req.user.rol !== 'Admin') return res.status(401).send('No eres admin');
+
 	const {idProducto, idCategoria} = req.params;
 
 	product_categories
 		.destroy({where: {productId: idProducto, categoryId: idCategoria}})
 		.then(() => res.sendStatus(200));
 });
-
-// router.get('/:idProducto/pics', (req, res) => {
-//
-// });
-//
-//
-// router.put('/:idProducto/pics', (req, res) => {
-//
-// });
 
 module.exports = router;
